requirements.txt
streamlit
pypdf
langchain-core
langchain_community
faiss-cpu
sentence-transformers
python-dotenv
langchain-groq 
huggingface_hub